@@ -1,4 +1,4 @@
-module.exports = async ({ actions, params, props }) => {
+export default async ({ actions, params, props }) => {
     const { Spinner: spinner, ActionSequence, chalk } = global.arcli;
     
     console.log('');
