@@ -1,7 +1,5 @@
-
-
-const chalk = require('chalk');
-const config = require('../config.json');
+import chalk from 'chalk';
+import config from '../config.json' assert { type: 'json' };
 
 /**
  * errorFunc(commandName, ...message) Function
@@ -10,7 +8,7 @@ const config = require('../config.json');
  * @param ...message Arguments passed to the console.log() function.
  * @return Boolean Returns false so that the function can be used as a validation result.
  */
-const error = (commandName, ...message) => {
+export const error = (commandName, ...message) => {
     console.log('');
 
     console.log(chalk.red('Error:'), commandName);
@@ -21,13 +19,12 @@ const error = (commandName, ...message) => {
     return false;
 };
 
-const message = (commandName, ...message) => {
+export const message = (commandName, ...message) => {
     console.log('');
-    console.log(chalk.cyan(String(config.prompt.prefix).trim()), commandName, ...message);
+    console.log(
+        chalk.cyan(String(config.prompt.prefix).trim()),
+        commandName,
+        ...message,
+    );
     console.log('');
-};
-
-module.exports = {
-    message,
-    error,
 };
